backend/requirements.txt
django
django_rest_framework
requests
jet